+++ src/os2datascanner/projects/admin/organizations/models/organization.py
@@ -11,6 +11,7 @@
 # OS2datascanner is developed by Magenta in collaboration with the OS2 public
 # sector open source network <https://os2.eu/>.
 #
+from django import forms
 from django.db import models
 from django.utils.text import slugify
 from django.utils.translation import gettext_lazy as _
@@ -43,6 +44,14 @@ def replace_nordics(name: str):
     return name
 
 
+class HourField(models.TimeField):
+    """ Allow the user to pick time of day in the format HH:MM. """
+
+    def formfield(self, **kwargs):
+        kwargs['widget'] = forms.TimeInput(format='%H:%M', attrs={'type': 'time', 'step': '3600'})
+        return super().formfield(**kwargs)
+
+
 @Broadcasted.register
 class Organization(Core_Organization):
     """ Core logic lives in the core_organizational_structure app.
@@ -64,6 +73,10 @@ class Organization(Core_Organization):
         CustomRule,
         related_name='organizations',
         verbose_name=_('system rules'),
+    )
+    synchronization_time = HourField(
+        default="17:00",
+        verbose_name=_('synchronization time'),
     )
 
     def save(self, *args, **kwargs):